backend/requirements.txt
fastapi>=0.111,<1.0
pydantic>=2.7,<3.0
uvicorn[standard]>=0.30,<1.0    
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0

